Crate authzen

Re-exports

pub use proc_macros;

pub use service_util;

Modules

actions

Standard actions which are useful across many applications.

decision_makers

Decision maker client implementations provided for convienence.

storage_backends

Storage backend client implementations provided for convienence.

transaction_caches

Macros

action

Generate a custom action which can be queried for authorization and performed on a storage backend.

Structs

Event

The unit of work in an authorization query, which will either be accepted or rejected by a decision maker.

TxCacheEntity

Enums

ActionError

Represents the possible sources of error when performing an action which requires authorization.

Traits

ActionType

Compile time information about an action.

AsStorage

Connects an object with its backend representation for a specific backend.

AuthorizationContext

Wraps all components of an action requiring authorization which can be expected to be carried around in an application context.

DecisionMaker

Represents a policy decision point (could be astracted over an in-process memory, a remote api, etc.) which is capable of making authorization decisions using the provided Event.

Identifiable

ObjectType

Compile time information about an object.

StorageAction

Encapsulates the actual performance of a specific action given a suitable client and input.

StorageBackend

Represents a unique backend on which actions are performed. An example would be postgres, mysql, or your own custom implementation of an API.

StorageClient

A client for communicating with a storage backend. Typically this should be implemented for connection or client implementations for that backend, e.g. diesel_async::AsyncPgConnection.

StorageObject

An object’s representation specific to a particular backend. Often this will only be implemented once unless an object is stored in multipled different backends.

TransactionCache

TransactionCacheAction

Derive Macros

AuthzObject

Context