Struct authzen::Event

pub struct Event<Subject, Action, Object, Input, Context = ()>where
    Action: ?Sized,
    Object: ?Sized,{
    pub subject: Subject,
    pub action: PhantomData<Action>,
    pub object: PhantomData<Object>,
    pub input: Input,
    pub context: Context,
}

The unit of work in an authorization query, which will either be accepted or rejected by a decision maker.

Fields
subject: Subject

the entity performing the action; typically links back to a user / account; could be represented with an account id, a jwt, or even more information if necessary

action: PhantomData<Action>

the action which the subject is trying to perform; typically this will be a struct dervied from the action macro (e.g. Create<MyObject>)

object: PhantomData<Object>

the object wich the subject is attempting to act upon; note that this parameter is only used as a specification of the object’s type and service it belongs to

input: Input

the data provided which uniquely identifies the object(s) being acted upon; the type used here can be anything which is recognized as valid input for the specific action (see StorageAction to see how actions specify their acceptable inputs)

context: Context

any additional data which should or must be provided in order to fulfill the authorization decision; use this for any data which is not referring to objects being acted on

Implementations

impl<Subject, Action, Object, Input, Context> Event<Subject, Action, Object, Input, Context>where Action: ?Sized, Object: ?Sized,

pub fn dissolve( self ) -> (Subject, PhantomData<Action>, PhantomData<Object>, Input, Context)

impl<Subject, Action, Object, Input, Context> Event<Subject, Action, Object, Input, Context>where Action: ?Sized, Object: ?Sized,

pub fn dissolve_mut( &mut self ) -> (&mut Subject, &mut PhantomData<Action>, &mut PhantomData<Object>, &mut Input, &mut Context)

impl<Subject, Action, Object, Input, Context> Event<Subject, Action, Object, Input, Context>where Action: ?Sized, Object: ?Sized,

pub fn dissolve_ref( &self ) -> (&Subject, &PhantomData<Action>, &PhantomData<Object>, &Input, &Context)

impl<Subject, Action, Object, Input, Context> Event<Subject, Action, Object, Input, Context>where Action: ?Sized, Object: ?Sized,

pub fn subject(&self) -> &Subject

pub fn input(&self) -> &Input

pub fn context(&self) -> &Context

impl<Subject, Action, Object, Input, Context> Event<Subject, Action, Object, Input, Context>where Action: ?Sized, Object: ?Sized,

pub fn builder( ) -> EventBuilder<((), (), (), (), ()), Subject, Action, Object, Input, Context>

Create a builder for building Event. On the builder, call .subject(...), .action(...), .object(...), .input(...), .context(...) to set the values of the fields. Finally, call .build() to create the instance of Event.